requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
numpy
xgboost
requests
pydantic
python-multipart
joblib
python-dotenv